pyproject.toml
[tool.poetry]
name = "django-crud-creator"
version = "0.1.0"
description = ""
authors = ["felipe_santos"]
readme = "README.md"
packages = [{include = "django_crud_creator"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0.1"
pymysql = "^1.1.0"
mysql-connector-python = "^8.0.33"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
